backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
python-multipart==0.0.9

httpx==0.27.0
pypdf==4.3.1
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==10.4.0

langchain==0.2.12
langchain-community==0.2.11
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1

transformers==4.43.3
torch>=2.3.1
huggingface_hub==0.24.5
langchain-huggingface==0.0.3

